make HEALTH message more graceful when `$BEASTHOST` cannot be resolved (#31)
* make HEALTH message more graceful when `$BEASTHOST` cannot be resolved

Fix for cosmetic issue that occurs when using a name for `$BEASTHOST` and this Beast host cannot be resolved. This occurs, for example, when `$BEASTHOST` is a container name, and that container isn't running. 

There's no need for this verbosity, which makes the health status hard to read: the status message itself already indicates that there is a connection issue.

Old behavior was to show the error output of `ss` across multiple lines in the HEALTH message; new behavior is to simply say it's in FAIL state.

* make linter happy